model-serving/pyproject.toml
[project]
name = "model-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ctranslate2>=4.6.0",
    "fastapi>=0.115.12",
    "faster-whisper>=1.1.1",
    "google-genai>=1.18.0",
    "google-generativeai>=0.8.5",
    "librosa>=0.11.0",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "torch>=2.7.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
